refactored contact context
